Timeline with database connection
WIP: Reading first timestamps
