+++ src/app/api/quality_matrix/models.py
@@ -1,4 +1,8 @@
+from typing import Any
+
 from pydantic import BaseModel, Field
+from sqlalchemy import JSON, Column, Integer
+from sqlalchemy.orm import as_declarative, declared_attr
 
 
 class ColumnOutputModel(BaseModel):
@@ -6,3 +10,19 @@ class ColumnOutputModel(BaseModel):
     columns: dict[str, float] = Field(
         description="The ratio of quality for the required columns."
     )
+
+
+@as_declarative()
+class Base:
+    id: Any
+    __name__: str
+
+    @declared_attr
+    def __tablename__(cls) -> str:
+        return cls.__name__.lower()
+
+
+class QualityMatrix(Base):
+    id = Column(Integer, primary_key=True, index=True)
+    timestamp = Column(Integer, nullable=False)
+    quality_matrix = Column(JSON, nullable=False)
